Add files via upload
Sales Trend Analysis of E-Commerce data using MS Excel based on data over last 4 years to find the products that yield in high profit returns and the ones that lead to loss. The project also includes suggestions that can help in better ROI (Return on Investment) for both the profit and loss making segments.